+++ packages/react-native-worklets/apple/worklets/apple/AssertJavaScriptQueue.h
@@ -3,12 +3,12 @@
 // Copied from RCTJSThreadManager.mm
 static NSString *const RCTJSThreadName = @"com.facebook.react.runtime.JavaScript";
 
-static BOOL IsJavaScriptQueue()
+static inline BOOL IsJavaScriptQueue()
 {
   return [NSThread.currentThread.name isEqualToString:RCTJSThreadName];
 }
 
-static void AssertJavaScriptQueue()
+static inline void AssertJavaScriptQueue()
 {
   react_native_assert(IsJavaScriptQueue() && "This function must be called on the JavaScript queue");
 }